fix: package name refactor
