+++ src/main/java/io/github/eisopux/diagnostics/core/CompilerRunner.java
@@ -1,4 +1,4 @@
-package org.eisopux.diagnostics.core;
+package io.github.eisopux.diagnostics.core;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,8 +7,8 @@ import javax.tools.*;
 
 /**
  * CompilerRunner sets up and executes a Java compilation task using the system Java compiler, while
- * integrating custom data {@link org.eisopux.diagnostics.core.Collector}(s) and a {@link
- * org.eisopux.diagnostics.core.Reporter} to process and format the collected output.
+ * integrating custom data {@link Collector}(s) and a {@link
+ * Reporter} to process and format the collected output.
  */
 public class CompilerRunner {
